+++ packages/react/src/components/selectable/selectable.template.tsx
@@ -14,6 +14,7 @@ export function selectableTemplate({
   invalid,
   describedById,
   checked,
+  indeterminate,
   disabled,
   onChange,
   info
@@ -29,6 +30,7 @@ export function selectableTemplate({
       disabled={disabled}
       required={required}
       checked={checked}
+      indeterminate={indeterminate}
       infoFixed={info?.fixed}
       dsoChange={(e: CustomEvent<Event>) => onChange(e.detail)}
     >